device tree: Make device tree the same as in android9
This patch modifies the device tree files to enable and disable
additional nodes. The goal is to make the device tree the same as in the
android9 BSP. The assumption is that then everything according to audio
should work.

How was this done? I pulled the running the device tree from the
android14 hardware and decompiled it to human readable dts files

    adb  pull /sys/firmware/fdt kernel-device-tree.android14.data.v2
    dtc -I dtb -O dts kernel-device-tree.android14.data.v2 > kernel-device-tree.android14.data.v2.dts

The "dts" then I compared with a decompiled dts file from the android9
kernel.